refactor: move main logic into app module
